I'm fixing the user input check in main.py.
trajectory.compose_database function asks for some arguments:

- peptide_length
- interval (between samples)
- start_from (index from which to start the frame count)

Topic of this commit is peptide_length argument.

peptide_length specify the number of aminoacids that are inside the peptides
in the trajectory.

His value should be an integer, or None

His the user is asked to insert is value when running main.py

If you leave it blank, it will be None, and morphoscanner will
infer the size adaptively from .gro file. This should be the preferred
option, because it makes morphoscanner conform to the gromacs output.
This is actually the only way (and still not completely implemented)
 to analyze trajectory in which there are peptide of different size.

If you insert an integer as peptide_length, it wil be passed as fixed
peptide size, so morphoscanner will group atoms in group
of the inserted length. This works only if all the peptides in the trajectory
have the same size.

Now, there is a bug in the contact counting.

Analyzing prod.gro I found that grouping fails if it does not
find the dict key for a contact ('Parallel' or 'Antiparallel')

error:

  File "/home/lillo/git_repo/sap_analysis/main.py", line 143, in <module>
    trj.get_data()

  File "/home/lillo/git_repo/sap_analysis/morphoscanner/trajectory.py", line 256, in get_data
    self.get_sense()

  File "/home/lillo/git_repo/sap_analysis/morphoscanner/trajectory.py", line 125, in get_sense
    parallel = len(self.frames[e]['frame_data'].groupby('sense').get_group('parallel'))

  File "/home/lillo/anaconda3/envs/pytorch/lib/python3.7/site-packages/pandas/core/groupby/groupby.py", line 686, in get_group
    raise KeyError(name)

KeyError: 'parallel'

I will fix this very soon, now I have go to the grocery shop