Working on new trajectory with multidimensional peptides recognition
Forgot to commit last time so check for change in this commit.

I was working on a frame first object for the library.

So first the frame object is created. inside the frame object there are peptide.
Giving different peptides dimension
(size or length...short for number of aminoacids in a peptide)
each peptide will compute his operation on distances and contact
as a single entity, and results for each frame will be saved
inside the peptide object. Here I need to spread already calculated
distance maps on other peptides, to not recompute distances
multiple times. If i compute distance between peptide 0 and
peptide 1, the resulting distance map should be spreaded to
peptide 0 and peptide 1.
I know that this will impact on memory, because the data will be
duplicated.
A solution to this will be to have a single container for
distance maps of a single frame.
Ther problem is that i am not able to think a solution
that enables me to not buplicate data, to not duplicate operation,
to enable fast data retrival and to parallelize operation
at the same time.

A better explanation for the actual situation is:

I have peptides of different length. I have to calculate
the distance between each peptide and all the others peptides.
I do this as a tensor operation, that is fast.
Tensor does not enable to contain data of different size
(at least not as a dense tensor, which i'm using),
for example if I have 10 peptides of 12 aminoacids each
i will have a tensor of shape (10,12,3) where the last
dimension (3) are the cartesian coordinate (x, y, z).
If I also have 20 peptides of length 16, i will have a
tensor of shape (20,16,3).
In the case these peptides are in the same dataset
what I do is to make a dictionary. My dictionary
wil search for peptide length, and then I create
a tensor for each peptide size. Following
the example data before, my dictionary will
have two keys, that are the peptide length ( 12 and 16)
and the key value will be a tensor containing all the peptide
of that size.
Then I take each peptide (peptide coordinate), that
is a tensor that in this case can be of shape (12,3)
or (16,3) and I broadcast this tensor for the
dimension of the tensor inside the dictionary.
Lets do the first operation of distance calculation with a
peptide of length 12.
When I compute the distance of my sigle peptide of shape (12,3)
with tensor in dict.get(12) I will broadcast my tensor
in a shape (10,12,3), and when i compute with the tensor in
dict.get(16) I will have a tensor of (20,12,3).
The results are two tensors, that contain the distance maps
obtained by calculating the euclidean distance of my first peptides
with all the others peptides.
I actually compute this operation for each peptide with the full
tensor of peptides to measure with.
This is not an optimized approach, because I calculate the same distance many times.
When i calculate distance between peptide 0 and all the other peptides
I save the distance maps.
When i calculate the distance between peptide 1 and all other peptides
I recalculate the distance between peptide 1 and peptide 0.
When I calculate distance between peptide 2 and all the other peptides
i recalculate the distances between peptide 2 and peptide 0, and
also the distance between peptide 2 and peptide 1, that I know already too.
This SHOULD be optimized.

Then I need to find a way to compute contact maps in a fast way, but this will come.

Anyway, maybe I can try sparse tensor to compute distances.