Fix readme. TODO
Thinking to the actual limitation of the library, we need a comprehensive approach
to the handling  of the initial atoms.
To enable the user to select her/his preferred atoms and residues from the trajectory
there are two ways:
- leveraging MDAnalsis (mda) capability, wrapping our function to data (coordinate and atom informations)
that are inside an mda.Universe() and develop an interface/adapter function to enable data retrival from the Universe().
- parse the Universe() to get all the object that we need, and create our own objects, that in this sceanrio
start from an atom object, that will contain data regarding the corresponding atom. This is generally
the line that I have followed in the development. This approache is faster and gives flexibility, but require effort
to mantain an efficient database. Parsing  is the step of the workflow that give us the data on which we perform
analysis, so parsing have to be precise.
In future an efficient interface to MDAnalysis will be an important step for Morphoscanner.